+++ packages/tools/f-vue-icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+v3.10.0
+------------------------------
+*May 12 2022*
+
+### Added
+- `PaymentVisaTransparentNarrowIcon` icon.
+
+
 v3.9.0
 ------------------------------
 *April 29 2022*